Revert "Optimize NBDFinder for complete mode analysis with publication-quality results and scientific rigor"
